smi: fix performance issue in OrderedDict
When loading large MIB values, __setitem__ is here a performance painful
point. Condition "if key not in self.__keys" performs a linear scan of a
list, and this list can be big. Checking is key is already in dict is
faster.

On a micro-benchmark (time to run SNMP queries on a device, after a
fresh startup) this simple patch move performances from 30s to only 3s